third_party/blink/web_tests/external/wpt/svg/animations/smil-clock-value-out-of-range-minutes-expected.txt

This is a testharness.js-based test.
[FAIL] SMIL clock values with out-of-range minutes or seconds should be treated as invalid
  assert_equals: dur="01:99:01" should be invalid (minutes > 59 in HH:MM:SS) expected 0 but got 1.2577297687530518
Harness: the test ran to completion.
